added rsvp module

+++ src/modules/events/person/person.service.ts
@@ -1,32 +1,60 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Person } from '../../../entities/events/event_person.schema';
-import { Model } from 'mongoose';
+import mongoose, { Model } from 'mongoose';
+import { Events } from '../../../entities/events/events.schema';
+import { ObjectId } from 'mongodb';
 
 @Injectable()
 export class PersonService {
   constructor(
     @InjectModel(Person.name) private readonly personModel: Model<Person>,
+    @InjectModel(Events.name) private readonly eventsModel: Model<Events>,
   ) {}
 
   async getAllPerson() {
     return this.personModel.find().lean();
   }
 
-  async createPerson(people) {
-    return await this.personModel.create(people);
+  async createPerson(person) {
+    return await this.personModel.create(person);
   }
 
   async viewPerson(id) {
-    return await this.personModel.findById(id);
+    const person = await this.personModel.findById(id).lean();
+    const event = await this.eventsModel
+      .findById(new ObjectId(person.event_name))
+      .lean();
+
+    return { ...person, event };
   }
 
-  async editPerson(id, people) {
-    console.log(id);
-    return await this.personModel.findByIdAndUpdate(id, people);
+  async editPerson(id, person) {
+    return await this.personModel.findByIdAndUpdate(id, person);
   }
 
   async deletePerson(id) {
-    return await this.personModel.findByIdAndDelete(id);
+    const person = await this.personModel.findById(id);
+    const deletedFromEvent = await this.eventsModel.updateOne(
+      {
+        _id: new ObjectId(person.event_name),
+      },
+      { $pull: { persons: id } },
+    );
+    if (deletedFromEvent) {
+      return await this.personModel.findByIdAndDelete(id);
+    }
+  }
+
+  async getAllEvents() {
+    return this.eventsModel.find({}).lean();
+  }
+
+  async savePersonToEvent(id, person) {
+    const saved = await this.eventsModel.updateOne(
+      { _id: id },
+      { $push: { persons: person._id } },
+    );
+    return saved;
   }
 }
